+++ docs/version_history.md
@@ -1,6 +1,6 @@
 ### v.0.6.5-beta
-"Lucky Furious AI Mutants"  
-date  
+"Fist of Fury"  
+February 12, 2022  
 
 * feat: luck (extremely good/bad, corresponding stats and accomplishments)
 * feat: learn other fighters' moves
@@ -23,6 +23,8 @@ date
 * refactor: genetic algorithm for fight AI training
 * feat: remove help in Fat Girl encounter
 * feat: in genetic algo, mutation only applies to children
+* refactor: player as a package
+* feat: train a new powerful genetic fight AI (pop=32 fights=160 n_gen=128 gen=84)
 
 
 ### v.0.6.4-beta